♻️ 깃헙이슈11번, 일관성있게 임포트
